Clarify that yabridgectl isn't a standalone thing
I'm just remembering that this was confusing to me the first time;
it was recommending a 'simpler' way of doing things through yabridgectl.  I didn't understand that that was a wrapper around the underlying yabridge.

Could you just package them together as a single 'yabridge' download? 
i.e. do you have users who want yabridge but not yabridgectl?