Readd accidentally removed forward declaration
This got removed in 74dc8225d1aabebf0c8d42a8559c8d25851f05e5 when I
wanted to rework all `*Impl` classes the same way I did with
`YaPluginFactoryImpl`. This ended up not being possible, but
accidentally also removed this forward declaration. With unity builds
this did not cause issues however, but with regular builds it might
depending on which files are changed.