:boom: Reimplement VST3 audio processing
In the same way as 50c25c1cf0d1846ae3717b987ebdfa555ac6d693 did it for
VST2 plugins. Input and output audio data is now stored in a shared
memory buffer instead of being sent over the sockets. This reduces the
bridging overhead to a minimum since copying data was the most expensive
operation we were doing and we now only need to copy the entire buffer
once per processing cycle.