:boom: Rewrite the usage section
I was being lazy so I was waiting for some other distros to package
yabridge before rewriting this, but I might as well do it now.
Technically speaking there's more text here than before, but I've made
sure to put the most important bits in the first sentence of list item.
As it turns out, some people just don't read past somewhere halfway
through the second sentence of a paragraph, regardless of long that
paragraph is. So a bulleted list is probably the best way to go here.