Update the Wine downgrade snippet for apt distros
This is now even more of a mess, because for some reason the codename
suffix changed again with Wine 6.17. To keep this compact, we now just
add the `-1` suffix for Wine 6.1 to Wine 6.17, and the `-2` suffix for
Wine 6.17 and above.

While we're add it, might as well add another variant to make it easier
to install an older version of `winehq-stable` or `winehq-devel`, if you
for some reason would want to do that.