Remove the wine-tkg reference for Ubuntu
This guide is somewhat outdated now, and the builds linked there tend to
cause a lot of problems. On Ubuntu there's no proper solution to use a
Wine build with fsync at the moment. Your best bet will likely be to use
the non-makepkg wine-tkg build script and to either set up your `PATH`
so that those Wine binaries get used or to build your own package based
on that.