Fix missing detached flag in Process::Handle moves
This is why C++'s move semantics are footguns. The result of this was
that grouped plugins would no longer be detached when they were
moved (in this case, into the group connect handler closure). Which in
turn caused those plugins to block until the plugin host process
terminated, which thus ended up blocking the host indefinitely.